add CanCollectRecord.Where
